agents/requirements.txt
uagents==0.20.1
python-dotenv==1.1.0
google-genai==1.14.0
httpx==0.28.1
